Split app into lib and main part
