+++ src/main.rs
@@ -1,33 +1,13 @@
-mod encryption;
-#[allow(dead_code)]
-mod error;
-mod file;
-mod meta;
-
-use arrayref::array_ref;
-use std::convert::TryInto;
+use clap::Parser;
+use file_encryptor::encryption::try_parse;
+use file_encryptor::{error, get_hash, try_decrypt, try_encrypt};
+use rpassword::prompt_password;
 use std::fs::remove_file;
+use std::io;
 use std::io::Write;
 use std::path::Path;
-use std::{io, iter};
 
-use rpassword::prompt_password;
-
-use rand::distributions::Alphanumeric;
-use rand::{thread_rng, Rng};
-
-use crate::encryption::{
-    append_meta, decrypt_file, encrypt_file, get_meta, try_parse,
-};
-use crate::error::ErrorKind::WrongPassword;
-use crate::file::OpenOrCreate;
-use clap::Parser;
-use file::GetFileDirectory;
-use sha2::{Digest, Sha256};
-
-extern crate core;
-extern crate log;
-
+// TODO: make mod app
 #[derive(Parser, Debug)]
 struct AppData {
     #[clap(help = "Path to the file", required = true)]
@@ -41,79 +21,6 @@ struct AppData {
 
     #[clap(long = "keep", help = "Do not delete original file")]
     pub keep_original: bool,
-}
-
-fn get_hash(key: &str) -> error::Result<[u8; 32]> {
-    let mut hasher = Sha256::new();
-    hasher.update(key.as_bytes());
-
-    let hashed_key: [u8; 32] = hasher
-        .finalize()
-        .as_slice()
-        .try_into()?;
-    Ok(hashed_key)
-}
-
-fn try_decrypt(
-    file_path: &Path,
-    hash_from_key: [u8; 32],
-    preview: bool,
-) -> error::Result<()> {
-    //target_file.seek(SeekFrom::Start(MAGIC_STRING.len() as u64))?;
-    let meta = get_meta(file_path)?;
-    let nonce = meta.nonce;
-
-    //   .\\filename.txt
-    //   ./source.\\filename
-    //   ./source/.\\filename
-
-    let filename = &meta.filename;
-    //let decrypt_file_path = &file_path.to_owned()
-    //    .with_file_name(filename);
-
-    let decrypt_file_path = file_path.file_dir()?.join(filename);
-
-    println!(
-        "decrypt_file_path: {:?}",
-        decrypt_file_path
-    );
-    decrypt_file(
-        file_path,
-        &decrypt_file_path,
-        meta.len(),
-        // 0,
-        &hash_from_key,
-        &nonce,
-        preview,
-    )
-    .map_err(|e| error::Error::new(WrongPassword, e))?;
-
-    Ok(())
-}
-
-fn try_encrypt(file_path: &Path, hash_from_key: [u8; 32]) -> error::Result<()> {
-    let target_file_path = &file_path.with_extension("enc");
-
-    let mut rng = thread_rng();
-    let rand_string = iter::repeat(())
-        .map(|()| rng.sample(Alphanumeric))
-        .map(u8::from)
-        .take(19)
-        .collect::<Vec<u8>>();
-
-    let nonce = array_ref![rand_string.as_slice(), 0, 19];
-    {
-        encrypt_file(
-            file_path,
-            target_file_path,
-            &hash_from_key,
-            nonce,
-        )
-        .map_err(|e| error::Error::new(WrongPassword, e))?;
-
-        append_meta(nonce, file_path, target_file_path)?;
-    }
-    Ok(())
 }
 
 fn main() -> error::Result<()> {
@@ -193,17 +100,4 @@ fn main() -> error::Result<()> {
     }
 
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use hex_literal::hex;
-
-    #[test]
-    fn get_hash() {
-        assert_eq!(
-            crate::get_hash("FM7348mwmw73t").unwrap(),
-            hex!("66afe59af310865bc544c9d7a19ded0b1f8e6a1e797c3a1215a33175cae4023c")
-        );
-    }
 }
